LayoutTests/js/dfg-get-closure-var-put-closure-var-interference-expected.txt

Tests that the DFG is appropriately conservative for may-aliased but not must-aliased PutClosureVar and GetClosureVar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS runIt() is 113
PASS successfullyParsed is true

TEST COMPLETE
